Removed RestConnection builders and added decorator pattern to RestConnection
