--- src/main/java/com/synopsys/integration/rest/connection/UnauthenticatedRestConnection.java
@@ -1,51 +1,0 @@
-/**
- * integration-rest
- *
- * Copyright (C) 2018 Black Duck Software, Inc.
- * http://www.blackducksoftware.com/
- *
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements. See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership. The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package com.synopsys.integration.rest.connection;
-
-import org.apache.http.client.config.RequestConfig;
-import org.apache.http.impl.client.HttpClientBuilder;
-
-import com.synopsys.integration.log.IntLogger;
-import com.synopsys.integration.rest.proxy.ProxyInfo;
-
-public class UnauthenticatedRestConnection extends RestConnection {
-    public UnauthenticatedRestConnection(final IntLogger logger, final ProxyInfo proxyInfo) {
-        super(logger, proxyInfo);
-    }
-
-    public UnauthenticatedRestConnection(final IntLogger logger, final int timeout, final ProxyInfo proxyInfo) {
-        super(logger, timeout, proxyInfo);
-    }
-
-    @Override
-    public void populateHttpClientBuilder(final HttpClientBuilder httpClientBuilder, final RequestConfig.Builder defaultRequestConfigBuilder) {
-        // no additional fields to populate
-    }
-
-    @Override
-    public void completeConnection() {
-        // nothing additional needed to connect
-    }
-
-}
