+++ src/test/groovy/com/synopsys/integration/rest/it/RestConnectionTestHelper.groovy
@@ -24,8 +24,7 @@ package com.synopsys.integration.rest.it
 
 import com.synopsys.integration.log.LogLevel
 import com.synopsys.integration.log.PrintStreamIntLogger
-import com.synopsys.integration.rest.connection.RestConnection
-import com.synopsys.integration.rest.connection.UnauthenticatedRestConnectionBuilder
+import com.synopsys.integration.rest.connection.*
 import com.synopsys.integration.rest.proxy.ProxyInfo
 import okhttp3.OkHttpClient
 import org.apache.commons.lang3.math.NumberUtils
@@ -34,17 +33,17 @@ import org.junit.Assert
 import java.util.logging.Level
 import java.util.logging.Logger
 
-public class RestConnectionTestHelper {
-    private Properties testProperties;
+class RestConnectionTestHelper {
+    private Properties testProperties
 
-    private final String serverUrl;
+    private final String serverUrl
 
-    public RestConnectionTestHelper() {
+    RestConnectionTestHelper() {
         initProperties()
         this.serverUrl = getProperty(TestingPropertyKey.TEST_HTTPS_SERVER_URL)
     }
 
-    public RestConnectionTestHelper(final String serverUrlPropertyName) {
+    RestConnectionTestHelper(final String serverUrlPropertyName) {
         initProperties()
         this.serverUrl = testProperties.getProperty(serverUrlPropertyName)
     }
@@ -53,7 +52,7 @@ public class RestConnectionTestHelper {
         Logger.getLogger(OkHttpClient.class.getName()).setLevel(Level.FINE)
         testProperties = new Properties()
         final ClassLoader classLoader = Thread.currentThread().getContextClassLoader()
-        InputStream is = null;
+        InputStream is = null
         try {
             is = classLoader.getResourceAsStream("test.properties")
             testProperties.load(is)
@@ -83,19 +82,19 @@ public class RestConnectionTestHelper {
         }
     }
 
-    public String getProperty(final TestingPropertyKey key) {
+    String getProperty(final TestingPropertyKey key) {
         return getProperty(key.toString())
     }
 
-    public String getProperty(final String key) {
+    String getProperty(final String key) {
         return testProperties.getProperty(key)
     }
 
-    public String getIntegrationServerUrlString() {
+    String getIntegrationServerUrlString() {
         return serverUrl
     }
 
-    public URL getIntegrationServerUrl() {
+    URL getIntegrationServerUrl() {
         URL url
         try {
             url = new URL(getIntegrationServerUrlString())
@@ -105,12 +104,12 @@ public class RestConnectionTestHelper {
         return url
     }
 
-    public int getTimeout() {
+    int getTimeout() {
         int timeout = NumberUtils.toInt(getProperty(TestingPropertyKey.TEST_TIMEOUT), 300)
         return timeout
     }
 
-    public File getFile(final String classpathResource) {
+    File getFile(final String classpathResource) {
         try {
             final URL url = Thread.currentThread().getContextClassLoader().getResource(classpathResource)
             final File file = new File(url.toURI().getPath())
@@ -121,23 +120,16 @@ public class RestConnectionTestHelper {
         }
     }
 
-    public RestConnection getRestConnection() {
-        return getRestConnection(LogLevel.DEBUG, ProxyInfo.NO_PROXY_INFO);
+    RestConnection getRestConnection() {
+        return getRestConnection(LogLevel.DEBUG, ProxyInfo.NO_PROXY_INFO)
     }
 
-    public RestConnection getRestConnection(final LogLevel logLevel) {
-        return getRestConnection(logLevel, ProxyInfo.NO_PROXY_INFO);
+    RestConnection getRestConnection(final LogLevel logLevel) {
+        return getRestConnection(logLevel, ProxyInfo.NO_PROXY_INFO)
     }
 
-    public RestConnection getRestConnection(final LogLevel logLevel, ProxyInfo proxyInfo) {
-        UnauthenticatedRestConnectionBuilder builder = new UnauthenticatedRestConnectionBuilder();
-        builder.logger = new PrintStreamIntLogger(System.out, logLevel);
-        builder.baseUrl = getIntegrationServerUrl()
-        builder.timeout = getTimeout()
-        builder.setProxyInfo(proxyInfo)
-        builder.setAlwaysTrustServerCertificate(true)
-        RestConnection restConnection = builder.build()
-        return restConnection
+    RestConnection getRestConnection(final LogLevel logLevel, ProxyInfo proxyInfo) {
+        return new BasicRestConnection(new PrintStreamIntLogger(System.out, logLevel), getTimeout(), true, proxyInfo)
     }
 
 }
